[wallet/symbol/mobile] task: restructure mosaic utils
