+++ wallet/symbol/mobile/src/screens/mosaic/utils/index.js
@@ -1,3 +1,4 @@
-export * from './mosaic-display';
 export * from './mosaic-duration';
+export * from './mosaic-supply';
+export * from './number-format';
 export * from './validators';
